requirements.txt
algoliasearch==1.9.2
algoliasearch-django==1.2.4
amqp==2.1.1
awsebcli==3.8.3
Babel==2.3.4
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.2
blessed==1.9.5
botocore==1.4.64
celery==4.0.0
cement==2.8.2
certifi==2016.9.26
cffi==1.7.0
colorama==0.3.7
cryptography==1.5
Django==1.10
django-cors-middleware==1.3.1
django-extensions==1.7.1
django-filter==0.13.0
djangorestframework==3.4.4
djangorestframework-camel-case==0.2.0
djangorestframework-jwt==1.8.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
drfdocs==0.0.11
enum34==1.1.6
flower==0.9.1
futures==3.0.5
gunicorn==19.6.0
idna==2.1
ipaddress==1.0.16
jmespath==0.9.0
kombu==4.0.0
Markdown==2.6.6
meld3==1.0.2
ndg-httpsclient==0.4.2
pathspec==0.3.4
psycopg2==2.6.2
pyasn1==0.1.9
pycparser==2.14
pydotplus==2.0.2
PyJWT==1.4.2
pyOpenSSL==16.1.0
pyparsing==2.1.10
python-dateutil==2.5.3
pytz==2016.7
PyYAML==3.12
redis==2.10.5
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
supervisor==3.3.1
texttable==0.8.5
tornado==4.2
vine==1.1.3
wcwidth==0.1.7
websocket-client==0.37.0
